+++ graphannis/src/annis/util/mod.rs
@@ -1,4 +1,5 @@
 pub mod quicksort;
+pub mod sortablecontainer;
 
 use graphannis_core::serializer::KeyVec;